Restructured components directory
